requirements.txt
asgiref==3.2.3
asn1crypto==1.3.0
beautifulsoup4==4.8.2
blinker==1.4
Brotli==1.0.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.42.0
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-developer-panel==0.1.2
django-filter==2.2.0
django-icons==1.1.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
Flask==1.1.1
futures==3.1.1
get==2019.4.13
goslate==1.5.1
gunicorn==19.7.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
idna==2.8
itsdangerous==1.1.0
Jinja2==3.0.0a1
kaitaistruct==0.8
ldap3==2.6.1
MarkupSafe==1.1.1
oauthlib==3.1.0
olefile==0.44
passlib==1.7.2
Pigments==1.6
Pillow==4.2.1
post==2019.4.13
protobuf==3.10.0
psycopg2==2.7.3.1
public==2019.4.13
publicsuffix2==2.20191221
pyasn1==0.4.8
pycparser==2.19
PyDictionary==1.5.2
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyperclip==1.7.0
PySocks==1.7.1
python3-openid==3.1.0
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
selenium==3.141.0
six==1.14.0
sortedcontainers==2.1.0
soupsieve==1.9.5
sqlparse==0.3.0
tornado==6.0.3
tweepy==3.8.0
urllib3==1.25.8
urwid==2.1.0
Werkzeug==1.0.0
whitenoise==5.1.0
wsproto==0.14.1
zstandard==0.12.0